+++ test/config.ts
@@ -4,13 +4,11 @@ import {
 } from '../src/constants';
 import { QualysIntegrationConfig } from '../src/types';
 
-export const config: QualysIntegrationConfig = {
+export const config = {
   qualysApiUrl:
     process.env.QUALYS_API_URL || 'https://qualysapi.qg3.apps.qualys.com',
   qualysPassword: process.env.QUALYS_PASSWORD || 'password',
   qualysUsername: process.env.QUALYS_USERNAME || 'upter3aw',
   minScannedSinceDays: DEFAULT_SCANNED_SINCE_DAYS,
-  minScannedSinceISODate: '2020-09-11T23:00:30Z',
   minFindingsSinceDays: DEFAULT_FINDINGS_SINCE_DAYS,
-  minFindingsSinceISODate: '2020-09-11T23:00:30Z',
-};
+} as QualysIntegrationConfig;